+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.2
 - Updated image cdns.
+- Added Eidolon#picture.
 # 1.0.1
 - Added StarRailCharacterBuild#imageUrl.
 - Removed SkillTreeNode#previousNodeIds, please use SkillTreeNode#previousNodeId instead.